requirements.txt
# YOLOv5 Face Mask Detection Requirements
torch==2.0.0
torchvision==0.15.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
mediapipe==0.10.0
matplotlib==3.7.1
tqdm==4.65.0
Pillow==9.5.0
PyYAML==6.0
scipy==1.10.1
pandas==2.0.2
seaborn==0.12.2

# YOLOv5 specific requirements
gitpython>=3.1.30
tensorboard>=2.4.1
thop>=0.1.1
psutil